requirements.txt
Django==2.2.1
djangorestframework==3.9.3
djangorestframework_simplejwt==4.3.0
django-cors-headers==3.0.1
psycopg2-binary==2.8.2
flake8==3.7.7
mixer==6.1.3
celery==4.3.0
redis==3.2.1
coverage==4.5.3
Pillow==6.0.0
